refactor: modify location
