+++ tomcat/src/main/java/org/apache/coyote/request/RequestHeader.java
@@ -3,8 +3,8 @@ package org.apache.coyote.request;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-import org.apache.catalina.controller.cookie.HttpCookie;
-import org.apache.catalina.requestMapping.session.Session;
+import org.apache.catalina.cookie.HttpCookie;
+import org.apache.catalina.session.Session;
 import org.apache.coyote.response.responseHeader.HttpHeaders;
 
 public class RequestHeader {
